+++ json_serializable/test/test_sources/test_sources.dart
@@ -17,7 +17,6 @@ class ConfigurationImplicitDefaults {
   generateToJsonFunction: true,
   includeIfNull: true,
   nullable: true,
-  useWrappers: false,
 )
 class ConfigurationExplicitDefaults {
   int field;